Footer Fix
Footer is now universal
